2 new mini games in fun.py
Added 2 mini games 
Dice game that works with command = {prefix}dice and the bot "Throws" dice with the player. 
Roulette game that works with command = {prefix}roulette {color} . Example: !roulette red. Then the bot spins the roulette with the 4 colors and if the color is same with players choise. Player wins.